requirements.txt
psutil==6.0.0
mysql-connector-python==9.1.0
streamlit==1.38.0
pandas==2.2.3
win10toast-click==0.9
python-dotenv==1.0.1
altair==5.5.0
